Accurate collisions that don't take mass into account

+++ CircleShape.h
@@ -9,7 +9,7 @@
 
 class CircleShape : public Shape {
 public:
-    float r;
+    double r;
 
     CircleShape(float r);
